// frontend/src/bindings/proto/tricked.ts
export interface HardwareMetrics {
    /**
     * @generated from protobuf field: float cpu_usage = 1
     */
    cpuUsage: number;
    /**
     * @generated from protobuf field: repeated float cpu_cores_usage = 2
     */
    cpuCoresUsage: number[];
    /**
     * @generated from protobuf field: float ram_usage_pct = 3
     */
    ramUsagePct: number;
    /**
     * @generated from protobuf field: float ram_used_mb = 4
     */
    ramUsedMb: number;
    /**
     * @generated from protobuf field: float gpu_util = 5
     */
    gpuUtil: number;
    /**
     * @generated from protobuf field: float vram_used_mb = 6
     */
    vramUsedMb: number;
    /**
     * @generated from protobuf field: float disk_usage_pct = 7
     */
    diskUsagePct: number;
    /**
     * @generated from protobuf field: float network_rx_mbps = 8
     */
    networkRxMbps: number;
    /**
     * @generated from protobuf field: float network_tx_mbps = 9
     */
    networkTxMbps: number;
    /**
     * @generated from protobuf field: float disk_read_mbps = 10
     */
    diskReadMbps: number;
    /**
     * @generated from protobuf field: float disk_write_mbps = 11
     */
    diskWriteMbps: number;
    /**
     * @generated from protobuf field: string machine_identifier = 12
     */
    machineIdentifier: string;
}
// ...
export const HardwareMetrics = new HardwareMetrics$Type();